# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This api is the backend api responsible for the whole Coo-e coordinator app.",
    "title": "Coordinator",
    "version": "",
    "x-stoplight": {
      "id": "haridasu/specs/coordinator"
    }
  },
  "host": "resources.coo-e.com:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/teeups/{teeup_uid}/when": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "gamePlanWhen": {
                  "createdAt": "2017-01-03T07:14:24.244Z",
                  "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                  "fromDate": "2016-12-29T08:43:52.000Z",
                  "id": "32aee301-cd33-4b18-8356-952c7d242ee6",
                  "teeupId": "1b546bc4-1981-4b05-afab-fbced8922cd6",
                  "thumbsDown": [
                    "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                  ],
                  "thumbsDownCount": 1,
                  "thumbsUp": [],
                  "thumbsUpCount": 0,
                  "toDate": "2016-12-29T09:34:07.000Z"
                },
                "status": 200,
                "whenSuggestions": [
                  {
                    "createdAt": "2017-01-03T07:14:24.244Z",
                    "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                    "fromDate": "2016-12-29T08:43:52.000Z",
                    "id": "32aee301-cd33-4b18-8356-952c7d242ee6",
                    "teeupId": "1b546bc4-1981-4b05-afab-fbced8922cd6",
                    "thumbsDown": [
                      "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                    ],
                    "thumbsDownCount": 1,
                    "thumbsUp": [],
                    "thumbsUpCount": 0,
                    "toDate": "2016-12-29T09:34:07.000Z"
                  },
                  {
                    "createdAt": "2017-01-03T09:38:13.486Z",
                    "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                    "fromDate": "2016-12-29T08:43:52.000Z",
                    "id": "909dd8b1-6855-4b62-b64c-f31e3a0914f1",
                    "teeupId": "1b546bc4-1981-4b05-afab-fbced8922cd6",
                    "thumbsDownCount": 0,
                    "thumbsUp": [
                      "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                    ],
                    "thumbsUpCount": 1,
                    "toDate": "2016-12-29T09:34:07.000Z"
                  }
                ]
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Invalid field parameters, takes boolean fields isPublic, join, invite, suggest, decideGamePlan, and modifyRow.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "Gets a list of when suggestions for a teeup and if theres a when in the gameplan then it will get that too",
        "summary": "Get a list of when suggestions for a Teeup"
      }
    },
...
    "/api/when": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "examples": {
              "application/json": {
                "message": "Successfully created suggestion!",
                "status": 201,
                "when": {
                  "createdAt": "2017-01-05T08:33:12.431Z",
                  "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                  "fromDate": "2016-12-29T08:43:52.000Z",
                  "id": "f65b51cd-93fd-4150-90d3-ad6b4f6c0f22",
                  "teeupId": "f003bfba-acf7-4c60-a99a-c476dce0a124",
                  "thumbsDownCount": 0,
                  "thumbsUp": [
                    "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                  ],
                  "thumbsUpCount": 1,
                  "toDate": "2016-12-29T09:34:07.000Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Teeup not found.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "Let's the user create a when suggestion\nthe params in the body is milliseconds",
        "summary": "Create a When Suggestion"
      }
    },
    "/api/when/{when_id}": {
      "delete": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Successfully retrieved when suggestion",
                "status": 200,
                "when": {
                  "createdAt": "2017-01-05T08:33:12.431Z",
                  "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                  "fromDate": "2016-12-29T08:43:52.000Z",
                  "id": "f65b51cd-93fd-4150-90d3-ad6b4f6c0f22",
                  "teeupId": "f003bfba-acf7-4c60-a99a-c476dce0a124",
                  "thumbsDownCount": 0,
                  "thumbsUp": [
                    "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                  ],
                  "thumbsUpCount": 1,
                  "toDate": "2016-12-29T09:34:07.000Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Unauthorized.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "Deletes a when suggestion and the details about it",
        "summary": "delete when suggestion"
      },
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Successfully retrieved when suggestion",
                "status": 200,
                "when": {
                  "createdAt": "2017-01-05T08:33:12.431Z",
                  "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                  "fromDate": "2016-12-29T08:43:52.000Z",
                  "id": "f65b51cd-93fd-4150-90d3-ad6b4f6c0f22",
                  "teeupId": "f003bfba-acf7-4c60-a99a-c476dce0a124",
                  "thumbsDownCount": 0,
                  "thumbsUp": [
                    "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                  ],
                  "thumbsUpCount": 1,
                  "toDate": "2016-12-29T09:34:07.000Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Unauthorized.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "Gets a when suggestion and the details about it",
        "summary": "Get when suggestion"
      }
    },
    "/api/when/{when_id}/dislike": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Successfully disliked!",
                "status": 200,
                "when": {
                  "createdAt": "2017-01-05T08:33:12.431Z",
                  "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                  "fromDate": "2016-12-29T08:43:52.000Z",
                  "id": "f65b51cd-93fd-4150-90d3-ad6b4f6c0f22",
                  "teeupId": "f003bfba-acf7-4c60-a99a-c476dce0a124",
                  "thumbsDown": [
                    "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                  ],
                  "thumbsDownCount": 1,
                  "thumbsUp": [],
                  "thumbsUpCount": 0,
                  "toDate": "2016-12-29T09:34:07.000Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Unauthorized.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Dislike a when suggestion"
      }
    },
    "/api/when/{when_id}/like": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Successfully liked!",
                "status": 200,
                "when": {
                  "createdAt": "2017-01-05T08:33:12.431Z",
                  "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                  "fromDate": "2016-12-29T08:43:52.000Z",
                  "id": "f65b51cd-93fd-4150-90d3-ad6b4f6c0f22",
                  "teeupId": "f003bfba-acf7-4c60-a99a-c476dce0a124",
                  "thumbsDown": [],
                  "thumbsDownCount": 0,
                  "thumbsUp": [
                    "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                  ],
                  "thumbsUpCount": 1,
                  "toDate": "2016-12-29T09:34:07.000Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Unauthorized.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Like a when suggestion"
      }
    },
    "/api/when/{when_id}/makeGamePlan": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Successfully updated gameplan",
                "status": 200,
                "when": {
                  "createdAt": "2017-01-05T08:33:12.431Z",
                  "createdBy": "a47537c7-958c-424c-81bb-d1d0612c0f9e",
                  "fromDate": "2016-12-29T08:43:52.000Z",
                  "id": "f65b51cd-93fd-4150-90d3-ad6b4f6c0f22",
                  "teeupId": "f003bfba-acf7-4c60-a99a-c476dce0a124",
                  "thumbsDown": [
                    "a47537c7-958c-424c-81bb-d1d0612c0f9e"
                  ],
                  "thumbsDownCount": 1,
                  "thumbsUp": [],
                  "thumbsUpCount": 0,
                  "toDate": "2016-12-29T09:34:07.000Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Unauthorized.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Make when the gamplan"
      }
    },
...
  },
...
}